Refactor and Rename
Changed the values of a few properties to be more uniform, and gave the base weapon the ability to always hit no matter what. This will obviously change next pass, but wanted to make sure all data was able to be passed through comfortably, allowing for quick fine tuning when I'm next free.