Weapon features branch (#23)
* Refactoring and Weapon class/asset

Refactored the following:
-Organized directories to fit the unity UI a bit better
-Organized the fields of a couple classes
-Refactored GetInput() to update existing struct rather than recreate it
-Added a new bool to the current ShipControllerInput struct to recognize when a weapon press is called (This may change to be more of an on-call event handler in the future)

Additionally, I added a ~very~ primitive weapon asset with crosshairs that attach to the player to show where their shots will be going. This particular portion is still WiP.

* Refactor and Rename

Changed the values of a few properties to be more uniform, and gave the base weapon the ability to always hit no matter what. This will obviously change next pass, but wanted to make sure all data was able to be passed through comfortably, allowing for quick fine tuning when I'm next free.

* First Pass Weapon System

Added a crude base weapon that is able to shoot, deal damage, and ultimately destroy an enemy. It can check for range with a set blast radius area, as well as a max range outwards.