Refactor for Future Features
Changed the way the ship communicates with weapons to be directly through the hardpoints, giving the hardpoints the ability to communicate to it's attachment. I created two enums (Size and Types) that can be used to determine how things can fit if needed, and the hardpoints can use the attachment through the inherited Use method (this is a rename of the FireWeapon method to allow for easier abstraction).