Attachments branch (#27)
* Hardpoints added

Addition of prototype hardpoints to the ship, as well as prefab to attach weapons too.

* Visual Changes

Found where to attach weapons/hardpoint to give them the same movement visuals as the rest of the ship.

* Start

* Update Game.unity

* Refactor for Future Features

Changed the way the ship communicates with weapons to be directly through the hardpoints, giving the hardpoints the ability to communicate to it's attachment. I created two enums (Size and Types) that can be used to determine how things can fit if needed, and the hardpoints can use the attachment through the inherited Use method (this is a rename of the FireWeapon method to allow for easier abstraction).

* Adjusted to fit guidelines

Adjusted syntax to fit the submission guidelines.